[PE] Pad section contents when rebuilding binary
Some PE have sections with section.VirtualSize < section.SizeOfRawData,
which is usually padded with zeroes. In that case, the PE parser only
read the lesser of the two to get only the useful content.

However, when rebuilding the binary as-is, we need to re-pad thoses
sections, otherwise the NT loader won't recognized it as a valid PE.