MachO::Binary::shift_command: fix exports of dyld (#1133)
* MachO::Binary::shift_command: fix shift of dyld exports

Despite the field name `address` it stores a file offset.

Co-authored-by: Richard Dzenis <rdzenis@whitecryption.com>

* MachO::Binary::shift_command: fix comparison in DyldExportsTrie update

`ExportInfo` is meant to be updated if `info.address()` is beyond the
point where shift happens; however, offset comparison was reversed.

Additionally, compare with `from_offset`, because despite the field name
`address` it stores a file offset.

Co-authored-by: Richard Dzenis <rdzenis@whitecryption.com>

* Add macho/test_exports_after_shift

This test checks that exports in dyld_info are shifted correctly.

* Add macho/test_chained_exports_after_shift

This test checks that exports in dyld_exports_trie
are shifted correctly.

---------

Co-authored-by: Peteris Ledins <peteris.ledins@zimperium.com>
Co-authored-by: Javier Lopez-Gomez <javier.lopez-gomez@zimperium.com>