Handle non-roman characters and extend OID To String support. (#1226)
* Remove upper character print restriction

* Extend OID support

Previous fallback for unknown OID was to print ??, now it falls back to the OID to String table.

* Update x509.cpp

Update variable names

* Return OID Numeric String if not found

After more testing, I confirmed that if the OID string is not found in either mbedtls or LIEF's table, it would return a non-human-readable string. This change will return the numeric string that could be investigated.

* Revert `OIDToString` changes and add tests

---------

Co-authored-by: Romain Thomas <me@romainthomas.fr>