Ensure that added DYN ELF sections are properly aligned
The in-page offset from the source segment is now kept, since changing
it may change the semantics of certain types of contents (e.g. ARM64
executable code relying on ADRP instructions).

Additionally, the virtual address of a segment is now properly aligned
in accordance with p_align instead of being hardcoded at the page size.

Co-authored-by: Thomas Wachter <thomas.wachter@huawei.com>
Fixes: d717340c9 ("Improve the logic of sections insertion in ELF binaries")